+++ CHANGELOG.md
@@ -1,8 +1,25 @@
 ## [Unreleased]
 
+FEATURES:
+
+- Added support for an `azure-event-hubs` data source `kind` on the `launchdarkly_destination` resource.
+
+ENHANCEMENTS:
+
+- Upgraded the Go version to 1.16.
+
+- Upgraded the LaunchDarkly SDK version to v5.1.0+.
+
+BUG FIXES:
+
+- Updated the destination `config` `Elem` type to `TypeString` and made the `config` field required. Added improved validation to check fields for different destination kinds.
+
 NOTES:
 
+- The `launchdarkly_destination` resource's `enabled` argument has been deprecated in favor of `on`. Please update your config to use `on` in order to maintain compatibility with future versions.
+
 - The `launchdarkly_webhook` resource's `enabled` argument has been deprecated in favor of `on`. Please update your config to use `on` in order to maintain compatibility with future versions.
+
 - The `launchdarkly_webhook` data source's `enabled` attribute has been deprecated in favor of `on`. Please update your all references of `enabled` to `on` in order to maintain compatibility with future versions.
 
 ## [1.5.1] (March 16, 2021)